move: CounterコンポーネントをUtilityディレクトリに移動
