+++ frontend/src/components/Buttons/LikeButton.vue
@@ -10,7 +10,7 @@
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
-import Counter from '@/components/Buttons/Counter.vue'
+import Counter from '@/components/Utility/Counter.vue'
 
 export default {
   props: {
