+++ frontend/src/components/Buttons/PowerWithoutButton.vue
@@ -12,7 +12,7 @@
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
-import Counter from '@/components/Buttons/Counter.vue'
+import Counter from '@/components/Utility/Counter.vue'
 
 export default {
   props: {
