+++ frontend/src/components/Utility/Counter.vue
